renamed to vk_error
